--- src/Background.js
@@ -1,19 +1,0 @@
-import styled from "styled-components";
-import ImgFundo from './img/fundoDaLoja.jpg'
-
-export const Fundo = styled.div `
-  background: url(${ImgFundo});
-  background-position: 30% 45%;
-  background-size: cover;
-  display: table;
-  width: 100%;
-  height: 100vh;
-
-`
-
-export const BotaoCrescenteDecrescente = styled.div`
-text-align: center;
-
-
-
-`
